+++ processing/segmenter/planktoscope/segmenter/__init__.py
@@ -49,23 +49,9 @@ import planktoscope.segmenter.encoder
 import planktoscope.segmenter.streamer
 import planktoscope.segmenter.ecotaxa
 
-
-################################################################################
-# Morphocut Libraries
-################################################################################
-# import morphocut
-# import morphocut.file
-# import morphocut.image
-# import morphocut.stat
-# import morphocut.stream
-# import morphocut.str
-# import morphocut.contrib.zooprocess
-
 ################################################################################
 # Other image processing Libraries
 ################################################################################
-import skimage.util
-import skimage.transform
 import skimage.measure
 import skimage.exposure
 import cv2